+++ src/Job/Job.h
@@ -7,7 +7,6 @@
 #include "include/cef_base.h"
 
 #include <string>
-#include <future>
 #include <functional>
 
 namespace cefpdf {
@@ -17,7 +16,18 @@ class Job : public CefBaseRefCounted
 {
 
 public:
-    typedef std::function<void(const std::string&, CefRefPtr<Job>)> Callback;
+    typedef std::function<void(CefRefPtr<Job>)> Callback;
+
+    enum struct Status {
+        PENDING,
+        LOADING,
+        PRINTING,
+        SUCCESS,
+        HTTP_ERROR,
+        ABORTED,
+        LOAD_ERROR,
+        PRINT_ERROR
+    };
 
     Job();
 
@@ -25,9 +35,9 @@ public:
         m_callback = callback;
     }
 
-    void ExecuteCallback(const std::string& reason) {
+    void ExecuteCallback() {
         if (m_callback != nullptr) {
-            m_callback(reason, this);
+            m_callback(this);
         }
     }
 
@@ -52,13 +62,21 @@ public:
     // Get prepared PDF setting for CEF
     CefPdfPrintSettings GetCefPdfPrintSettings() const;
 
+    Status GetStatus() {
+        return m_status;
+    }
+
+    void SetStatus(Status status) {
+        m_status = status;
+    }
+
 private:
     CefString m_outputPath;
     PageSize m_pageSize;
     PageOrientation m_pageOrientation;
     PageMargin m_pageMargin;
     bool m_backgrounds;
-    std::promise<std::string> m_promise;
+    Status m_status;
     Callback m_callback;
 
     // Include the default reference counting implementation.