+++ src/Server/Session.cpp
@@ -10,6 +10,8 @@
 
 #include <iostream>
 #include <string>
+#include <sstream>
+#include <ostream>
 #include <functional>
 #include <regex>
 
@@ -26,10 +28,42 @@ Session::Session(
     m_socket(std::move(socket))
 {}
 
-void Session::Read()
+void Session::ReadHeaders()
 {
-    m_socket.async_read_some(
-        asio::buffer(m_buffer),
+    asio::async_read_until(
+        m_socket,
+        m_buffer,
+        http::crlf + http::crlf,
+        std::bind(
+            &Session::OnRead,
+            this,
+            std::placeholders::_1,
+            std::placeholders::_2
+        )
+    );
+}
+
+void Session::ReadExactly(std::size_t bytes)
+{
+    asio::async_read(
+        m_socket,
+        m_buffer,
+        asio::transfer_exactly(bytes),
+        std::bind(
+            &Session::OnRead,
+            this,
+            std::placeholders::_1,
+            std::placeholders::_2
+        )
+    );
+}
+
+void Session::ReadAll()
+{
+    asio::async_read(
+        m_socket,
+        m_buffer,
+        asio::transfer_all(),
         std::bind(
             &Session::OnRead,
             this,
@@ -41,16 +75,23 @@ void Session::Read()
 
 void Session::Write()
 {
-    asio::async_write(
-        m_socket,
-        ResponseToBuffers(),
-        std::bind(
-            &Session::OnWrite,
-            this,
-            std::placeholders::_1,
-            std::placeholders::_2
-        )
-    );
+    std::error_code error;
+    asio::streambuf buffer;
+    std::ostream os(&buffer);
+
+    m_response.WriteToStream(os);
+
+    asio::write(m_socket, buffer, error);
+
+    if (!error) {
+        // Initiate graceful connection closure.
+        asio::error_code ignored_ec;
+        m_socket.shutdown(asio::ip::tcp::socket::shutdown_both, ignored_ec);
+    }
+
+    if (error != asio::error::operation_aborted) {
+        m_sessionManager->Stop(this);
+    }
 }
 
 void Session::OnRead(std::error_code ec, std::size_t bytes_transferred)
@@ -59,83 +100,128 @@ void Session::OnRead(std::error_code ec, std::size_t bytes_transferred)
         return;
     }
 
-    m_requestData += std::string(m_buffer.data(), bytes_transferred);
+    if (m_request.method.empty()) {
+        ParseRequestHeaders();
 
-    if (m_socket.available() > 0) {
-        Read();
-    } else {
-        ParseRequest();
-        if (!Handle()) {
+        if (m_request.method == "POST") {
+            if (stringsEqual(m_request.expect, "100-continue")) {
+                Write100Continue();
+            }
+
+            if (m_request.contentLength > 0) {
+                std::size_t bytes_to_read = m_request.contentLength - m_request.content.size();
+                if (bytes_to_read > 0) {
+                    ReadExactly(bytes_to_read);
+                } else {
+                    HandlePOST();
+                }
+            } else {
+                // No content length: read until connection closes
+                ReadAll();
+            }
+
+        } else if (m_request.method == "GET") {
+            HandleGET();
+        } else {
+            m_response.SetStatus(http::statuses::badMethod);
             Write();
         }
+    } else {
+        m_request.content += FetchBuffer();
+        HandlePOST();
     }
 }
 
-void Session::OnWrite(std::error_code ec, std::size_t bytes_transferred)
+std::string Session::FetchBuffer()
 {
-    if (!ec) {
-        // Initiate graceful connection closure.
-        asio::error_code ignored_ec;
-        m_socket.shutdown(asio::ip::tcp::socket::shutdown_both, ignored_ec);
-    }
+    std::ostringstream ss;
+    ss << &m_buffer;
 
-    if (ec != asio::error::operation_aborted) {
-        m_sessionManager->Stop(this);
-    }
+    return ss.str();
 }
 
-void Session::ParseRequest()
+void Session::Write100Continue()
 {
+    std::error_code error;
+    asio::streambuf buffer;
+    std::ostream os(&buffer);
+
+    os << http::protocol << " " << http::statuses::cont << http::crlf << http::crlf;
+
+    asio::write(m_socket, buffer, error);
+}
+
+void Session::ParseRequestHeaders()
+{
+    auto requestData = FetchBuffer();
+
+    // Separate headers section from request content
+    std::string headerData;
+    auto sepIndex = requestData.find(http::crlf + http::crlf);
+
+    if (sepIndex == std::string::npos) {
+        headerData = requestData;
+    } else {
+        headerData = requestData.substr(0, sepIndex + 2);
+        m_request.content = requestData.substr(sepIndex + 4, requestData.size() - sepIndex + 4);
+    }
+
     // Parse status line
-    std::regex re("^([A-Z]+) (\\S+) HTTP/(.+)\r\n");
+    std::regex re("^([A-Z]+) (\\S+) HTTP/(.+)" + http::crlf);
     std::smatch match;
 
-    if (std::regex_search(m_requestData, match, re)) {
+    if (std::regex_search(headerData, match, re)) {
         m_request.method = match[1];
         m_request.url = match[2];
         m_request.version = match[3];
     }
 
-    // Separate headers section from request content
-    std::string headerData;
-    auto sepIndex = m_requestData.find(http::crlf + http::crlf);
-
-    if (sepIndex == std::string::npos) {
-        headerData = m_requestData;
-    } else {
-        headerData = m_requestData.substr(0, sepIndex);
-        m_request.content = m_requestData.substr(sepIndex + 4, m_requestData.size() - sepIndex + 4);
-    }
+    m_request.contentLength = 0;
 
     // Collect request headers
-    std::regex re2("(.+): (.+)\r\n");
+    std::regex re2("(.+): (.+)" + http::crlf);
     std::string::const_iterator it, end;
     it = match[0].second;
-    end = m_requestData.end();
+    end = headerData.end();
 
     while (std::regex_search(it, end, match, re2)) {
         m_request.headers.push_back({match[1], match[2]});
+
+        if (stringsEqual(match[1], http::headers::encoding)) {
+            m_request.transferEncoding = match[2];
+        } else if (stringsEqual(match[1], http::headers::length)) {
+            m_request.contentLength = std::stoi(match[2]);
+        } else if (stringsEqual(match[1], http::headers::expect)) {
+            m_request.expect = match[2];
+        } else if (stringsEqual(match[1], http::headers::location)) {
+            m_request.location = match[2];
+        }
+
         it = match[0].second;
     }
 }
 
-bool Session::Handle()
+void Session::HandleGET()
 {
     m_response.SetStatus(http::statuses::ok);
     m_response.SetHeader(http::headers::date, formatDate());
-
-    if (!(m_request.method == "GET" || m_request.method == "POST")) {
-        m_response.SetStatus(http::statuses::badMethod);
-        return false;
-    }
 
     if (m_request.url == "/" || m_request.url == "/about") {
         m_response.SetContent(
             "{\"status\":\"ok\",\"version\":\"" + cefpdf::constants::version + "\"}",
             "application/json"
         );
-        return false;
+    } else {
+        m_response.SetStatus(http::statuses::notFound);
     }
+
+    Write();
+}
+
+void Session::HandlePOST()
+{
+    m_response.SetStatus(http::statuses::ok);
+    m_response.SetHeader(http::headers::date, formatDate());
 
     // Parse url path
     std::regex re("([^/]+\\.pdf)($|[^\\w])", std::regex_constants::icase);
@@ -143,92 +229,44 @@ bool Session::Handle()
 
     if (!std::regex_search(m_request.url, match, re)) {
         m_response.SetStatus(http::statuses::notFound);
-        return false;
+        Write();
+        return;
     }
 
     std::string fileName = match[1];
 
     m_response.SetHeader(http::headers::disposition, "inline; filename=\"" + fileName + "\"");
 
-    std::string location;
-
-    for (auto h: m_request.headers) {
-        if (h.name.compare(http::headers::location) == 0) {
-            location = h.value;
-            break;
-        }
-    }
-
     CefRefPtr<cefpdf::job::Job> job;
 
-    if (location.empty()) {
-        if (m_request.method != "POST") {
-            m_response.SetStatus(http::statuses::badMethod);
-            return false;
-        }
-
-        if (m_request.content.empty()) {
-            m_response.SetStatus(http::statuses::badRequest);
-            return false;
-        }
-
+    if (m_request.location.empty()) {
         job = new cefpdf::job::Local(m_request.content);
     } else {
-        job = new cefpdf::job::Remote(location);
+        job = new cefpdf::job::Remote(m_request.location);
     }
 
     job->SetCallback(std::bind(
         &Session::OnResolve,
         this,
-        std::placeholders::_1,
-        std::placeholders::_2
+        std::placeholders::_1
     ));
 
     CefPostTask(TID_UI, base::Bind(&cefpdf::Client::AddJob, m_client, job));
-    return true;
 }
 
-void Session::OnResolve(const std::string& result, CefRefPtr<cefpdf::job::Job> job)
+void Session::OnResolve(CefRefPtr<cefpdf::job::Job> job)
 {
     if (!m_socket.is_open()) {
         return;
     }
 
-    if (result == "success") {
+    if (job->GetStatus() == cefpdf::job::Job::Status::SUCCESS) {
         m_response.SetContent(loadTempFile(job->GetOutputPath()), "application/pdf");
     } else {
         m_response.SetStatus(http::statuses::error);
     }
 
-    std::error_code error;
-
-    asio::write(
-        m_socket,
-        ResponseToBuffers(),
-        error
-    );
-
-    OnWrite(error, 0);
-}
-
-std::vector<asio::const_buffer> Session::ResponseToBuffers()
-{
-    std::vector<asio::const_buffer> buffers;
-
-    buffers.push_back(asio::buffer(m_response.status));
-    buffers.push_back(asio::buffer(http::crlf));
-
-    for (std::size_t i = 0; i < m_response.headers.size(); ++i) {
-        buffers.push_back(asio::buffer(m_response.headers[i].name));
-        buffers.push_back(asio::buffer(http::hsep));
-        buffers.push_back(asio::buffer(m_response.headers[i].value));
-        buffers.push_back(asio::buffer(http::crlf));
-    }
-
-    buffers.push_back(asio::buffer(http::crlf));
-    buffers.push_back(asio::buffer(m_response.content));
-
-    return buffers;
+    Write();
 }
 
 } // namespace server